refactor(web): rename form-canvas tool to form-designer
Rename the 2D form tool form-canvas → form-designer end to end: tool dir,
id, component (FormDesignerTool), web-core toolRegistry entry + the package's
relatedTools, the tools/messages aggregators + registry spec, and the display
name (Form Canvas/表單畫布 → Form Designer/表單設計器). Route is registry-driven,
so it becomes /tools/form-designer automatically; the old slug 404s.
