--- apps/web/src/tools/form-canvas/index.ts
@@ -1,5 +1,0 @@
-import type { ToolModule } from "@/tools/types";
-
-import { FormCanvasTool } from "./ui";
-
-export const tool: ToolModule = { id: "form-canvas", Component: FormCanvasTool };
