--- apps/web/src/tools/form-canvas/messages.ts
@@ -1,21 +1,0 @@
-import type { LocaleMessages } from "@/tools/types";
-
-export const messages: LocaleMessages = {
-  en: {
-    Tools: {
-      "form-canvas": {
-        title: "Form Canvas",
-        description:
-          "A 2D-canvas form builder — drag fields on a 12-column grid, group them, and preview live (prototype).",
-      },
-    },
-  },
-  "zh-TW": {
-    Tools: {
-      "form-canvas": {
-        title: "表單畫布",
-        description: "2D 畫布式表單建構器：在 12 欄網格上拖放欄位、分組，並即時預覽（原型）。",
-      },
-    },
-  },
-};
