+++ apps/web/src/tools/form-designer/inspector/labels.spec.tsx
