+++ apps/web/src/tools/form-designer/inspector/misc-sections.spec.tsx
