+++ apps/web/src/tools/form-designer/inspector/settings-panel.spec.tsx
