+++ apps/web/src/tools/messages.ts
@@ -3,7 +3,7 @@ import type { LocaleMessages } from "./types";
 import { messages as dataFilterBuilder } from "./data-filter-builder/messages";
 import { messages as dataFilterTester } from "./data-filter-tester/messages";
 import { messages as formBuilder } from "./form-builder/messages";
-import { messages as formCanvas } from "./form-canvas/messages";
+import { messages as formDesigner } from "./form-designer/messages";
 import { messages as esClientDemo } from "./es-client-demo/messages";
 import { messages as esQueryBuilder } from "./es-query-builder/messages";
 import { messages as jsonbQueryBuilder } from "./jsonb-query-builder/messages";
@@ -32,5 +32,5 @@ export const toolMessages: LocaleMessages[] = [
   esClientDemo,
   pgFilterBuilder,
   formBuilder,
-  formCanvas,
+  formDesigner,
 ];
